hosts: Use FQDN provided via $SET_FQDN or $SET_HOSTNAME
Both SET_HOSTNAME and SET_FQDN can contain a full qualified domain
name. The motivation for adding another variable for the same thing is
to allow users specify the system's FQDN when instanciating a template
and at the same time have a safe fallback to `SET_HOSTNAME = $NAME` when
the FQDN variable is not provided. This is basically a workaround for
inflexiblity of ONe's context variables (you cannot define a default
value of a variable).